feat(cicd): Sets up pnpm via github action (#166)
## Type of PR (check all applicable)

- [x] Feature
- [ ] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Performance Improvement
- [ ] Test Update
- [ ] Other (please describe)

## Description

This pull request includes changes to the
`.github/actions/blt/action.yml` file to update the setup process for
Node.js and pnpm. The most important changes involve the removal of
Corepack setup steps and the addition of a specific action to set up
pnpm.

Improvements to the setup process:

*
[`.github/actions/blt/action.yml`](diffhunk://#diff-5add2b4469a302a6125b9cc24c7f0990bbdcad55d294536b6a0f1505ae783ce3L23-L35):
Replaced the Corepack setup steps with the `pnpm/action-setup@v4` action
to streamline the pnpm setup process.
*
[`.github/actions/blt/action.yml`](diffhunk://#diff-5add2b4469a302a6125b9cc24c7f0990bbdcad55d294536b6a0f1505ae783ce3L23-L35):
Simplified the Node.js setup by removing the `fetch-depth` parameter and
ensuring the cache is set for pnpm.