feat(web): Add magic link email provider for login (#443)
## Type of PR (check all applicable)

- [ ] Feature

## Description

This adds the possibility to login via magic links.

To test it locally, you can use a tool like Mailcatcher, see:
https://mailcatcher.me/. You can also run it inside Docker, see:
https://hub.docker.com/r/dockage/mailcatcher/

To enable login via magic links, you need to add the following to your
local apps/web/.env file:

```
# Nodemailer for magic email links
AUTH_EMAIL_SERVER=smtp://localhost:1025
AUTH_EMAIL_FROM=noreply@localhost
```

Remarks:
* The feature is not complete yet. We still need a registration page
with a Captcha to block potential bots. That is why you should not use
this on a public facing website until that has been completed. It can
however be used for local testing with multiple accounts.
* The texts in the email sent are currently hardcoded. They have been
added to the i18n package, but it needs to be figured out how to obtain
the locale from the email being sent.