feat(web): Annotations (frontend) (#637)
## Type of PR (check all applicable)

- [x] Feature
- [ ] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Performance Improvement
- [ ] Test Update
- [ ] Other (please describe)

## Description

This features the UI for Comments & Flags without bulk actions. Since
the backend still needs to be changed, this is a bit work in progress.

## Related Issues

- Related to #153

## Testing Instructions

This cannot be tested separately without the backend being changed. To
enable demo state, please make the following changes to in
apps/web/hooks/experiment/useExperimentData/useExperimentData.ts:
- uncomment lines 7 and 215-223
- comment out line 224

## Changes Made

- Added the displaying of comments in each row with a on hover card to
show the comments and flags. It also contains the option to add or
delete comments or flags for a single row.

## Checklist

- [ ] I have added/updated tests for my changes
- [ ] My code follows the project's style guidelines
- [ ] I have updated the documentation accordingly
- [ ] I have tested these changes locally
- [ ] My changes generate no new warnings
- [ ] I have reviewed my own code