feat(backend): Notify added user (#766)
## Type of PR (check all applicable)

- [x] Feature
- [ ] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Performance Improvement
- [ ] Test Update
- [ ] Chore
- [ ] Other (please describe)

## Description

When a user is added to an experiment, each added member will get a
service email. This email shows to which experiment the user has been
added, by whom and a link to the project.

## Related Issues

- Closes #720

## Changes Made

- Adds email module to backend to send service emails
- Added notification message which gets called when a member is added to
an experiment

## Checklist

- [ ] I have added/updated tests for my changes
- [ ] My code follows the project's style guidelines
- [ ] I have updated the documentation accordingly
- [ ] I have tested these changes locally
- [ ] My changes generate no new warnings
- [ ] I have reviewed my own code

## Additional Notes

This PR is work in progress. The following things still need to be done:

- Fix the following error in unit tests which occurs as soon as it tries
to send emails:

Error: connect ECONNREFUSED 127.0.0.1:1025
 ❯ TCPConnectWrap.afterConnect [as oncomplete] node:net:1634:16
 ❯ TCPConnectWrap.callbackTrampoline node:internal/async_hooks:130:17

We can start the Docker container MailCatcher while running the unit
tests in the backend.

- Check the texts and formatting for the email
- Add unit tests to experiment-members.controller.spec.ts for the added
logic
- Add unit tests for email.adapter.ts
- Add unit tests for notifications.service.ts
- Prepare the hosting environment for service emails

You need to add the following to your local backend/.env file:

```
# Nodemailer for service emails
EMAIL_BASE_URL=http://localhost:3000
EMAIL_SERVER=smtp://localhost:1025
EMAIL_FROM=noreply@localhost
```

---------

Co-authored-by: petar-INFO <petar@info.nl>