chore(cicd): minor tweaks around tf format checks and app deploys, adds option to do a full deploy manually (#909)
## Type of PR (check all applicable)

- [ ] Feature
- [ ] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Performance Improvement
- [ ] Test Update
- [x] Chore
- [ ] Other (please describe)

## Description

This pull request introduces improvements to the deployment workflows
and the OpenTofu format check action. The main updates add a
`deploy_all` input to multiple workflows, allowing all applications to
be deployed or released regardless of detected changes, and make the
OpenTofu format check more configurable and strict in pull request
checks.